report missing label once: closes https://github.com/metanorma/metanorma-iso/issues/285
